quiz creation + repartition data
